pyproject.toml
[project]
name = "yt2navidrome"
version = "0.0.1"
description = "CLI tool to download YT videos and playlists with metadata required for Navidrome"
authors = [
    {name = "Prehistoic", email = "lacotematthieu@gmail.com"}
]
readme = "README.md"
license = "BSD-3-Clause"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
  "click>=8.1.0,<9.0",
  "click-option-group>=0.5.9,<1.0",
  "python-dotenv>=1.2.1,<2.0",
  "pyfiglet>=1.0.4,<2.0",
  "yt-dlp[default]>=2025.11.12",
  "PyYAML>=6.0.3,<7.0"
]

[dependency-groups]
dev = [
  "pytest>=7.2.0,<8.0",
  "pytest-cov>=4.0.0,<5.0",
  "deptry>=0.16.2,<1.0",
  "mypy>=1.5.1,<2.0",
  "types-yt-dlp>=2025.11.12",
  "pre-commit>=3.4.0,<4.0",
  "tox>=4.11.1,<5.0",
]

[project.urls]
repository = "https://github.com/prehistoic/yt2navidrome"
documentation = "https://prehistoic.github.io/yt2navidrome/"

[project.scripts]
yt2navidrome = "yt2navidrome.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["yt2navidrome"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["yt2navidrome"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
